cache CI activity data across tab switches
