+++ plugins/self-service/src/components/GitRepositories/RepositoriesCIActivityTab.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { Box, Link, Paper, Typography } from '@material-ui/core';
 import { useTheme } from '@material-ui/core/styles';
 import Autocomplete from '@material-ui/lab/Autocomplete';
@@ -26,163 +26,11 @@ import {
   useCollectionsStyles,
   useTableWrapperStyles,
 } from '../CollectionsCatalog/styles';
-import { formatTimeAgo, getSourceUrl } from '../CollectionsCatalog/utils';
-import { getProjectDisplayName } from './scmUtils';
-import { BatchItem, buildBatchItems } from './ciBatchUtils';
-import { CI_BATCH_CHUNK_SIZE, CI_PARALLEL_LIMIT } from './constants';
+import { formatTimeAgo } from '../CollectionsCatalog/utils';
+import { CIActivityRow, STATUS_LABELS } from './ciActivityUtils';
+import { ciActivityCache, CIActivityCacheState } from './ciActivityCache';
 
-export type CIActivityRow = {
-  id: string;
-  status:
-    | 'success'
-    | 'failure'
-    | 'cancelled'
-    | 'in_progress'
-    | 'queued'
-    | 'skipped'
-    | 'unknown';
-  project: string;
-  projectUrl?: string;
-  event: string;
-  eventDisplay: string;
-  trigger: string;
-  time: string;
-  runUrl?: string;
-};
-
-function normalizeGitLabStatus(
-  s: string | undefined | null,
-): CIActivityRow['status'] {
-  const status = (s ?? 'unknown').toLowerCase();
-  const map: Record<string, CIActivityRow['status']> = {
-    success: 'success',
-    failed: 'failure',
-    canceled: 'cancelled',
-    cancelled: 'cancelled',
-    running: 'in_progress',
-    pending: 'in_progress',
-    skipped: 'skipped',
-  };
-  return map[status] ?? 'unknown';
-}
-
-const STATUS_LABELS: Record<CIActivityRow['status'], string> = {
-  success: 'Success',
-  failure: 'Failure',
-  cancelled: 'Cancelled',
-  in_progress: 'In Progress',
-  queued: 'Queued',
-  skipped: 'Skipped',
-  unknown: 'Unknown',
-};
-
-function normalizeGitHubStatus(raw: string): CIActivityRow['status'] {
-  const valid: CIActivityRow['status'][] = [
-    'success',
-    'failure',
-    'cancelled',
-    'in_progress',
-    'queued',
-    'skipped',
-  ];
-  const s = raw.toLowerCase();
-  return valid.includes(s as CIActivityRow['status'])
-    ? (s as CIActivityRow['status'])
-    : 'unknown';
-}
-
-function parseGitHubRuns(
-  key: string,
-  data: unknown,
-  entity: Entity,
-): CIActivityRow[] {
-  const ghData = data as Record<string, unknown> | undefined;
-  const runs = (
-    Array.isArray(ghData?.workflow_runs) ? ghData.workflow_runs : []
-  ) as Record<string, unknown>[];
-  const repoUrl = getSourceUrl(entity);
-  const baseUrl = (repoUrl ?? '').replace(/\.git$/i, '');
-  const projectUrl = baseUrl ? `${baseUrl}/actions` : undefined;
-  const projectName = getProjectDisplayName(entity);
-
-  return runs.map((r: Record<string, unknown>) => {
-    const run = r as Record<string, string | number | undefined>;
-    return {
-      id: `gh-${key}-${run.id}`,
-      status: normalizeGitHubStatus(
-        String(run.conclusion ?? run.status ?? 'unknown'),
-      ),
-      project: projectName,
-      projectUrl,
-      event: String(run.name ?? 'Workflow'),
-      eventDisplay: `${run.name ?? 'Workflow'} #${run.run_number ?? run.id}`,
-      trigger: String(run.event ?? 'unknown').replaceAll('_', ' '),
-      time: String(run.created_at ?? ''),
-      runUrl: run.html_url ? String(run.html_url) : undefined,
-    };
-  });
-}
-
-function parseGitLabPipelines(
-  key: string,
-  data: unknown,
-  entity: Entity,
-): CIActivityRow[] {
-  const pipelines = (Array.isArray(data) ? data : []) as Record<
-    string,
-    unknown
-  >[];
-  const repoUrl = getSourceUrl(entity);
-  const baseUrl = (repoUrl ?? '').replace(/\.git$/i, '');
-  const projectUrl = baseUrl ? `${baseUrl}/-/pipelines` : undefined;
-  const projectName = getProjectDisplayName(entity);
-
-  return pipelines.map((p: Record<string, unknown>) => {
-    const pipeline = p as Record<string, string | number | undefined>;
-    return {
-      id: `gl-${key}-${pipeline.id}`,
-      status: normalizeGitLabStatus(
-        pipeline.status !== null && pipeline.status !== undefined
-          ? String(pipeline.status)
-          : undefined,
-      ),
-      project: projectName,
-      projectUrl,
-      event: 'Pipeline',
-      eventDisplay: `Pipeline #${pipeline.id}`,
-      trigger: String(pipeline.source ?? 'unknown').replaceAll('_', ' '),
-      time: String(pipeline.created_at ?? ''),
-      runUrl: pipeline.web_url ? String(pipeline.web_url) : undefined,
-    };
-  });
-}
-
-function buildRowsFromResults(
-  results: Record<
-    string,
-    { status: number; data: unknown } | { error: string }
-  >,
-  entityMap: Map<string, { entity: Entity; provider: string }>,
-): CIActivityRow[] {
-  const allRows: CIActivityRow[] = [];
-
-  for (const [key, result] of Object.entries(results)) {
-    if ('error' in result) continue;
-    const entry = entityMap.get(key);
-    if (!entry) continue;
-
-    if (entry.provider === 'github') {
-      allRows.push(...parseGitHubRuns(key, result.data, entry.entity));
-    } else if (entry.provider === 'gitlab') {
-      allRows.push(...parseGitLabPipelines(key, result.data, entry.entity));
-    }
-  }
-
-  allRows.sort(
-    (a, b) => new Date(b.time).getTime() - new Date(a.time).getTime(),
-  );
-  return allRows.slice(0, 150);
-}
+export type { CIActivityRow };
 
 export interface RepositoriesCIActivityTabProps {
   filterByEntity?: Entity | null;
@@ -200,88 +48,44 @@ export const RepositoriesCIActivityTab = ({
   const discoveryApi = useApi(discoveryApiRef);
   const fetchApi = useApi(fetchApiRef);
 
-  const [loading, setLoading] = useState(true);
-  const [fetchingMore, setFetchingMore] = useState(false);
-  const [rows, setRows] = useState<CIActivityRow[]>([]);
-  const [error, setError] = useState<string | null>(null);
+  const [cacheState, setCacheState] = useState<CIActivityCacheState | null>(
+    ciActivityCache.getState(),
+  );
   const [statusFilter, setStatusFilter] = useState<string>('All');
   const [triggerFilter, setTriggerFilter] = useState<string>('All');
 
-  const fetchActivity = useCallback(async () => {
-    setLoading(true);
-    setFetchingMore(false);
-    setError(null);
-    setRows([]);
-
-    try {
-      let entities: Entity[];
-      if (filterByEntity) {
-        entities = [filterByEntity];
-      } else if (cachedEntities && cachedEntities.length > 0) {
-        entities = cachedEntities;
-      } else {
-        const response = await catalogApi.getEntities({
-          filter: [{ kind: 'Component', 'spec.type': 'git-repository' }],
-        });
-        entities = Array.isArray(response) ? response : (response?.items ?? []);
-      }
-
-      const { items: batchItems, entityMap } = buildBatchItems(entities, 15);
-
-      if (batchItems.length === 0) {
-        setRows([]);
-        setLoading(false);
-        return;
-      }
-
-      const catalogBase = await discoveryApi.getBaseUrl('catalog');
-      const batchUrl = `${catalogBase}/ansible/git/ci-activity`;
-      const allResults: Record<
-        string,
-        { status: number; data: unknown } | { error: string }
-      > = {};
-
-      const chunks: BatchItem[][] = [];
-      for (let i = 0; i < batchItems.length; i += CI_BATCH_CHUNK_SIZE) {
-        chunks.push(batchItems.slice(i, i + CI_BATCH_CHUNK_SIZE));
-      }
-
-      const fetchChunk = async (chunk: BatchItem[]) => {
-        const res = await fetchApi.fetch(batchUrl, {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          credentials: 'include',
-          body: JSON.stringify({ items: chunk }),
-        });
-
-        if (!res.ok) {
-          throw new Error(`Batch CI activity request failed: ${res.status}`);
-        }
-
-        return res.json();
-      };
-
-      setLoading(false);
-      setFetchingMore(true);
-
-      for (let i = 0; i < chunks.length; i += CI_PARALLEL_LIMIT) {
-        const batch = chunks.slice(i, i + CI_PARALLEL_LIMIT);
-        const results = await Promise.all(batch.map(fetchChunk));
-        results.forEach(body => Object.assign(allResults, body.results));
-        setRows(buildRowsFromResults(allResults, entityMap));
-      }
-      setFetchingMore(false);
-    } catch (e) {
-      setError(e instanceof Error ? e.message : 'Failed to load CI activity');
-    } finally {
-      setLoading(false);
-      setFetchingMore(false);
-    }
-  }, [catalogApi, discoveryApi, fetchApi, filterByEntity, cachedEntities]);
-
   useEffect(() => {
-    fetchActivity();
-  }, [fetchActivity]);
+    const unsubscribe = ciActivityCache.subscribe(setCacheState);
+
+    if (filterByEntity) {
+      ciActivityCache.startLoading([filterByEntity], discoveryApi, fetchApi);
+    } else if (cachedEntities && cachedEntities.length > 0) {
+      ciActivityCache.startLoading(cachedEntities, discoveryApi, fetchApi);
+    } else {
+      catalogApi
+        .getEntities({
+          filter: [{ kind: 'Component', 'spec.type': 'git-repository' }],
+        })
+        .then(response => {
+          const items = Array.isArray(response)
+            ? response
+            : (response?.items ?? []);
+          ciActivityCache.startLoading(items, discoveryApi, fetchApi);
+        })
+        .catch(e => {
+          ciActivityCache.setError(
+            e instanceof Error ? e.message : 'Failed to load entities',
+          );
+        });
+    }
+
+    return unsubscribe;
+  }, [discoveryApi, fetchApi, catalogApi, filterByEntity, cachedEntities]);
+
+  const rows = useMemo(() => cacheState?.rows ?? [], [cacheState?.rows]);
+  const loading = !cacheState;
+  const fetchingMore = cacheState?.fetchingMore ?? false;
+  const error = cacheState?.error ?? null;
 
   const statusOptions = useMemo(() => {
     const statuses = Array.from(new Set(rows.map(r => r.status))).sort((a, b) =>
